# backend/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Server for SAC App",
        "title": "SAC API",
        "contact": {
            "name": "David Oduneye and Garrett Ladley",
            "email": "generatesac@gmail.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Logs in a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logs in a user",
                "operationId": "login-user",
                "parameters": [
                    {
                        "description": "Login Body",
                        "name": "loginBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginUserResponseBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utilities.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
...
        "/auth/me": {
            "get": {
                "description": "Returns the current user associated with an auth session",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Retrieve the current user given an auth session",
                "operationId": "get-current-user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "get": {
                "description": "Refreshes a user's access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refreshes a user's access token",
                "operationId": "refresh-user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utilities.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
        "/auth/update-password/:userID": {
            "post": {
                "description": "Updates a user's password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Updates a user's password",
                "operationId": "update-password",
                "parameters": [
                    {
                        "description": "User Body",
                        "name": "userBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdatePasswordRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utilities.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errors.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "models.Club": {
            "type": "object",
            "required": [
                "application_link",
                "description",
                "is_recruiting",
                "name",
                "num_members",
                "preview",
                "recruitment_cycle",
                "recruitment_type"
            ],
            "properties": {
                "application_link": {
                    "type": "string",
                    "maxLength": 255
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "description": {
                    "description": "MongoDB URL",
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "is_recruiting": {
                    "type": "boolean"
                },
                "logo": {
                    "description": "S3 URL",
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "num_members": {
                    "type": "integer",
                    "minimum": 1
                },
                "preview": {
                    "type": "string",
                    "maxLength": 255
                },
                "recruitment_cycle": {
                    "maxLength": 255,
                    "enum": [
                        "fall",
                        "spring",
                        "fallSpring",
                        "always"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.RecruitmentCycle"
                        }
                    ]
                },
                "recruitment_type": {
                    "maxLength": 255,
                    "enum": [
                        "unrestricted",
                        "tryout",
                        "application"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.RecruitmentType"
                        }
                    ]
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
        "models.College": {
            "type": "string",
            "enum": [
                "CAMD",
                "DMSB",
                "KCCS",
                "CE",
                "BCHS",
                "SL",
                "CPS",
                "CS",
                "CSSH"
            ],
            "x-enum-comments": {
                "BCHS": "Bouvé College of Health Sciences",
                "CAMD": "College of Arts, Media and Design",
                "CE": "College of Engineering",
                "CPS": "College of Professional Studies",
                "CS": "College of Science",
                "CSSH": "College of Social Sciences and Humanities",
                "DMSB": "D'Amore-McKim School of Business",
                "KCCS": "Khoury College of Computer Sciences",
                "SL": "School of Law"
            },
            "x-enum-varnames": [
                "CAMD",
                "DMSB",
                "KCCS",
                "CE",
                "BCHS",
                "SL",
                "CPS",
                "CS",
                "CSSH"
            ]
        },
...
        "models.LoginUserResponseBody": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 8
                }
            }
        },
...
        "models.RecruitmentCycle": {
            "type": "string",
            "enum": [
                "fall",
                "spring",
                "fallSpring",
                "always"
            ],
            "x-enum-varnames": [
                "Fall",
                "Spring",
                "FallSpring",
                "Always"
            ]
        },
        "models.RecruitmentType": {
            "type": "string",
            "enum": [
                "unrestricted",
                "tryout",
                "application"
            ],
            "x-enum-varnames": [
                "Unrestricted",
                "Tryout",
                "Application"
            ]
        },
...
        "models.Tag": {
            "type": "object",
            "required": [
                "category_id",
                "name"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
...
        "models.UpdatePasswordRequestBody": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "required": [
                "college",
                "email",
                "first_name",
                "last_name",
                "nuid",
                "role",
                "year"
            ],
            "properties": {
                "clubs_followed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Club"
                    }
                },
                "college": {
                    "maxLength": 255,
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.College"
                        }
                    ]
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "nuid": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "super",
                        "student"
                    ]
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "year": {
                    "maximum": 6,
                    "minimum": 1,
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Year"
                        }
                    ]
                }
            }
        },
        "models.Year": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6
            ],
            "x-enum-varnames": [
                "First",
                "Second",
                "Third",
                "Fourth",
                "Fifth",
                "Graduate"
            ]
        },
        "utilities.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}